feat(tier1): widget store unification, dashboard badge parity, cost --days, R8 split
Batch of Tier-1 backlog items (#13 #40 #26 #20):

- 1a (#13): Move app-group id + metrics-key + read path into
  MeterBarShared.SharedMetricsStore; delete the widget's duplicate
  SharedDataStore; app-side store keeps the write path + widget reloads.
  StorageKeys re-exports the shared key. Contract tests pin the constants.
- 1b (#40): Shared ProviderStatusBadges component (resets-available +
  extra-usage pill) used by the popover AND dashboard provider cards.
- 1c (#40): Extract CodexCliUsageResponse.toUsageMetrics() pure mapping;
  fixture unit tests.
- 1d (#26): meterbar cost --days N — pure CostSummary.dailyCostWindow filter
  over cached daily rows (no rescan) with a truncation notice; unit-tested.
- 1e (#20): Delete stale MeterBar/Info.plist (build uses
  GENERATE_INFOPLIST_FILE); drop the stale SPM exclude.
- 1f: Split UsageDashboardView.swift (1539->450) into DashboardCard,
  DashboardProviderCards, DashboardCostCards, DailyUsageChart. Pure moves.
