+++ .agents/SESSIONS/2026-07-03.md
@@ -1,0 +1,98 @@
+# Sessions: 2026-07-03
+
+**Summary:** Tier 1 backlog batch (#13 #40 #26 #20)
+
+---
+
+## Session 1: Tier 1 small-scoped fixes (1a–1f)
+
+**Duration:** ~1 session
+**Status:** Complete
+
+### What was done
+
+Implemented all six Tier-1 items from `.agents/docs/ISSUE_BACKLOG.md` as one PR, TDD where logic was involved.
+
+- **1a — Widget `SharedDataStore` unification (rest of #13).** Added `SharedMetricsStore`
+  to `MeterBarShared` owning the app-group identifier, metrics file name, container/file
+  URLs, and the tolerant read path. Deleted the widget's private duplicate `SharedDataStore`
+  class; the widget now calls `SharedMetricsStore.loadMetrics()`. App-side `SharedDataStore`
+  keeps the write path + `reloadWidgetTimelines()` and delegates its read + location to the
+  shared type. `StorageKeys.cachedUsageMetrics` re-exports `SharedMetricsStore.metricsKey`
+  so the app/widget/CLI can't drift. Extended `CachedMetricsContractTests` to pin the shared
+  constants and the reader's decode path.
+- **1b — Dashboard badge parity (rest of #40).** New shared `ProviderStatusBadges` component
+  (`MeterBar/Views/Components/`) renders the "N resets available" + "Extra usage" pill,
+  `.compact` for the popover and `.regular` for the dashboard. Wired into
+  `PopoverProviderStatusCard`, `ProviderOverviewStatusCard`, and `ProviderLimitsCard` — no
+  copy-paste. Removed the duplicated `resetCreditsLabel` static from `MenuBarView`.
+- **1c — Split `CodexCliLocalService.fetchUsageMetrics()` (rest of #40).** Extracted the
+  response→`UsageMetrics` mapping into a pure `CodexCliUsageResponse.toUsageMetrics()`;
+  `fetchUsageMetrics()` now just decodes + calls it. Fixture-driven unit tests
+  (`CodexUsageMappingTests`) cover paid/free accounts and a missing secondary window.
+- **1d — `meterbar cost --days N` (rest of #26).** Added a pure `CostSummary.dailyCostWindow`
+  filter over the cached daily rows (no rescan) + `DailyCostWindow`/`ProviderDailyTotal`
+  types; wired `@Option days` into the CLI `Cost` subcommand with a truncation notice when
+  the cache covers fewer days than requested. Unit-tested the filter (`CostWindowTests`).
+- **1e — Deleted stale `MeterBar/Info.plist` (rest of #20).** Build uses
+  `GENERATE_INFOPLIST_FILE=YES`; the checked-in file hardcoded 1.0. Verified the only real
+  reference was the SPM `exclude` (removed it); the widget's Info.plist and the CLI's runtime
+  bundle-plist path are untouched.
+- **1f — Finished R8: split `UsageDashboardView.swift`** (1539→450 lines) into
+  `DashboardCard.swift`, `DashboardProviderCards.swift`, `DashboardCostCards.swift`, and
+  `DailyUsageChart.swift`. Pure moves; the only edits were dropping file-scope `private` on
+  moved top-level decls so they stay module-internal.
+
+### Files changed
+
+- `Packages/MeterBarShared/Sources/MeterBarShared/SharedMetricsStore.swift` - new shared store/constants
+- `MeterBar/Models/StorageKeys.swift` - re-export shared metrics key
+- `MeterBar/Services/SharedDataStore.swift` - delegate read + location to shared store
+- `MeterBarWidget/UsageWidget.swift` - drop duplicate store, use shared reader
+- `MeterBar/Services/CodexCliLocalService.swift` - extract `toUsageMetrics()`
+- `MeterBar/Models/TokenCost.swift` - `dailyCostWindow` + `DailyCostWindow`/`ProviderDailyTotal`
+- `MeterBarCLI/Sources/MeterBarCLI.swift` - `cost --days N`
+- `MeterBar/Views/Components/ProviderStatusBadges.swift` - new shared badge row
+- `MeterBar/Views/MenuBarView.swift` - use shared badges
+- `MeterBar/Views/UsageDashboardView.swift` - trimmed to window controller + main view; badges added
+- `MeterBar/Views/{DashboardCard,DashboardProviderCards,DashboardCostCards,DailyUsageChart}.swift` - R8 split
+- `MeterBar/Info.plist` - deleted (stale)
+- `Package.swift` - drop the stale `Info.plist` exclude
+- `MeterBarTests/{CachedMetricsContractTests,CodexUsageMappingTests,CostWindowTests}.swift` - tests
+
+### Decisions
+
+- **Decision:** Shared read store lives in `MeterBarShared` as `SharedMetricsStore`, not a
+  renamed `StorageKeys`.
+  - **Context:** The app already has an app-only `StorageKeys` (Claude account keys, etc.).
+  - **Rationale:** Avoids a name collision when the app imports `MeterBarShared`; keeps the
+    app's UserDefaults keys out of the cross-target shared surface.
+- **Decision:** `cost --days` reports only what daily rows carry (input/output/cache-read +
+  cost), not cache-creation tokens or session counts.
+  - **Context:** `DailyTokenUsage` doesn't store those; the dated data is the daily rows.
+  - **Rationale:** Honest, rescan-free windowing; a truncation notice fires when the cache
+    spans fewer days than requested rather than silently under-reporting.
+
+### Mistakes and fixes
+
+- **Mistake:** First pass at the R8 split sliced type ranges by declaration line, so
+  `DashboardDateFormat`'s leading doc comment landed at the end of the previous file.
+- **Fix:** Re-sliced with comment-aware start boundaries; verified the moved line multiset is
+  byte-identical to the original (modulo added imports and dropped `private`).
+- **Prevention:** When splitting Swift files, treat a type's block as starting at its leading
+  comment/attribute lines, not the `struct`/`enum` keyword line.
+
+### Verification
+
+- Local: only Command Line Tools installed (no XCTest/SourceKit), so per policy full
+  `swift test` + SwiftLint run in CI. Locally: built `MeterBarShared` (`swift build`, passes);
+  verified split content preservation, brace/paren balance, and no consecutive blank lines.
+
+### Next steps
+
+- [ ] CI: confirm `swift test` + SwiftLint `--strict` green on the PR.
+- [ ] #20 stays open — optional local `scripts/package.sh` mirroring `release.yml` not done.
+
+---
+
+**Total sessions today:** 1
