+++ MeterBar/Services/CodexCliLocalService.swift
@@ -127,57 +127,10 @@ class CodexCliLocalService: ObservableObject {
                 self.subscriptionType = usageResponse.planType
             }
 
-            // Check if rate limits exist (free accounts have null rate_limit)
-            guard let rateLimit = usageResponse.rateLimit else {
-                // Return empty metrics for free accounts
-                return UsageMetrics(
-                    service: .codexCli,
-                    sessionLimit: nil,
-                    weeklyLimit: nil,
-                    codeReviewLimit: nil,
-                    extraUsage: usageResponse.extraUsageStatus,
-                    resetCreditsAvailable: usageResponse.resetCreditsAvailable
-                )
-            }
-
-            // Map the response to UsageMetrics
-            // Primary window (5 hours = 18000 seconds) = session limit
-            let primaryWindow = rateLimit.primaryWindow
-            let sessionLimit = UsageLimit(
-                used: primaryWindow.usedPercent,
-                total: 100.0,
-                resetTime: Date(timeIntervalSince1970: Double(primaryWindow.resetAt)),
-                windowSeconds: TimeInterval(primaryWindow.limitWindowSeconds)
-            )
-
-            // Secondary window (7 days = 604800 seconds) = weekly limit
-            let secondaryWindow = rateLimit.secondaryWindow
-            let weeklyLimit = UsageLimit(
-                used: secondaryWindow?.usedPercent ?? 0.0,
-                total: 100.0,
-                resetTime: secondaryWindow.map { Date(timeIntervalSince1970: Double($0.resetAt)) } ?? Date(),
-                windowSeconds: secondaryWindow.map { TimeInterval($0.limitWindowSeconds) }
-            )
-
-            // Code review rate limit (7 days window) = code review limit
-            var codeReviewLimit: UsageLimit?
-            if let codeReviewPrimary = usageResponse.codeReviewRateLimit?.primaryWindow {
-                codeReviewLimit = UsageLimit(
-                    used: codeReviewPrimary.usedPercent,
-                    total: 100.0,
-                    resetTime: Date(timeIntervalSince1970: Double(codeReviewPrimary.resetAt)),
-                    windowSeconds: TimeInterval(codeReviewPrimary.limitWindowSeconds)
-                )
-            }
-
-            return UsageMetrics(
-                service: .codexCli,
-                sessionLimit: sessionLimit,
-                weeklyLimit: weeklyLimit,
-                codeReviewLimit: codeReviewLimit,
-                extraUsage: usageResponse.extraUsageStatus,
-                resetCreditsAvailable: usageResponse.resetCreditsAvailable
-            )
+            // Pure response→UsageMetrics mapping (window math, code-review limit,
+            // extra-usage + reset-credit derivation) lives on the response type
+            // so it's unit-testable with fixture JSON, no network.
+            return usageResponse.toUsageMetrics()
         } catch {
             let serviceError = ServiceSupport.serviceError(from: error)
             AppLog.usage.error("Codex usage fetch failed: \(serviceError.localizedDescription)")
@@ -271,6 +224,69 @@ struct CodexCliUsageResponse: Codable {
             detail += " · cap \(ExtraUsageStatus.formatAmount(limit))"
         }
         return detail
+    }
+}
+
+// MARK: - Response → UsageMetrics mapping
+
+extension CodexCliUsageResponse {
+    /// Pure mapping from the decoded Codex usage payload to the shared
+    /// `UsageMetrics`. Extracted from `fetchUsageMetrics()` so the window/limit
+    /// derivation is unit-testable with fixture JSON, no network.
+    ///
+    /// Windows map as: primary (5h) → session, secondary (7d) → weekly,
+    /// code-review primary (7d) → code-review limit. Free accounts (null
+    /// `rate_limit`) carry no windows, only the extra-usage/reset-credit signals.
+    func toUsageMetrics() -> UsageMetrics {
+        guard let rateLimit else {
+            // Free accounts have a null rate_limit — no quota windows to report.
+            return UsageMetrics(
+                service: .codexCli,
+                sessionLimit: nil,
+                weeklyLimit: nil,
+                codeReviewLimit: nil,
+                extraUsage: extraUsageStatus,
+                resetCreditsAvailable: resetCreditsAvailable
+            )
+        }
+
+        // Primary window (5 hours = 18000 seconds) = session limit
+        let primaryWindow = rateLimit.primaryWindow
+        let sessionLimit = UsageLimit(
+            used: primaryWindow.usedPercent,
+            total: 100.0,
+            resetTime: Date(timeIntervalSince1970: Double(primaryWindow.resetAt)),
+            windowSeconds: TimeInterval(primaryWindow.limitWindowSeconds)
+        )
+
+        // Secondary window (7 days = 604800 seconds) = weekly limit
+        let secondaryWindow = rateLimit.secondaryWindow
+        let weeklyLimit = UsageLimit(
+            used: secondaryWindow?.usedPercent ?? 0.0,
+            total: 100.0,
+            resetTime: secondaryWindow.map { Date(timeIntervalSince1970: Double($0.resetAt)) } ?? Date(),
+            windowSeconds: secondaryWindow.map { TimeInterval($0.limitWindowSeconds) }
+        )
+
+        // Code review rate limit (7 days window) = code review limit
+        var codeReviewLimit: UsageLimit?
+        if let codeReviewPrimary = codeReviewRateLimit?.primaryWindow {
+            codeReviewLimit = UsageLimit(
+                used: codeReviewPrimary.usedPercent,
+                total: 100.0,
+                resetTime: Date(timeIntervalSince1970: Double(codeReviewPrimary.resetAt)),
+                windowSeconds: TimeInterval(codeReviewPrimary.limitWindowSeconds)
+            )
+        }
+
+        return UsageMetrics(
+            service: .codexCli,
+            sessionLimit: sessionLimit,
+            weeklyLimit: weeklyLimit,
+            codeReviewLimit: codeReviewLimit,
+            extraUsage: extraUsageStatus,
+            resetCreditsAvailable: resetCreditsAvailable
+        )
     }
 }
